glfs: allow to load the device when the sizes disagree
Before this change:
The old code will just fail the tcmu device creating, but on the
resize operation, say the backend file expand succees but if the
config change on one tcmu-runner node fails. In this case, after
the node rebooted and retoring it will always fail to load the
tcmu device.

And we will have no chance to do any recovery of it, like to rerun
the resize again.

With this change:
It will always allow the tcmu device have a different size value
with the backend file when creating, and then glfs handler will
update the new size to kernel.

But the size attribute in saveconfig.json file still need to be
updated by runner the `gluster-block resize` command or `targetcli
saveconfig` command later.

Signed-off-by: Xiubo Li <xiubli@redhat.com>